We now only search when the search is confirmed
Previously, we used to search for every times letter, which not only wasn't helpful because the app literally couldn't refresh fast enough for this to even be noticeable, we also caused significant amounts of API calls to that open-source API that is thankfully operated for free, which is not a good practice. This way, we keep our API usage in check, ensuring that the developer of the API isn't absolutely demolished by requests. It also means the phone has to process less JSON, which will probably make it happy, especially as we will continue to expand the app and add features like on-the-fly unit conversions that happen while the JSON is being parsed. The principle is: don't underestimate the power of a device, but never overestimate its battery life - the lesser power draw, the better. We also save band with by not immediately starting to download big images.