Implemented Image Caching
This will save images you have been looking at to a cache such that they will load immediately next time. This is helpful because RCDB has some really high-resolution images that are quite large.

The implementation is only barebones so far but it took way too long to get this operational. For the moment, we have an unlimited amount of chacheable images meaning your storage will eventually be filled up. Also, the progress view and the view for "no image found" aren't ready yet. But they exist and they work and that's all that makes me a very happy person right now.