Visual adjustments
This commit aims to make the application more presentable in demos by changing some minor things. This is done to show a demo to the team behind RCDB such that I can obtain their permission to use their platform's data for this application. This will also mean the repo will from now on be public such that I can share a link to the team.